# Fix layer norm overflow issue on GPU