Summary:
Fix layer norm overflow issue on GPU

Performance benchmark on V100 machine:
input size = [32768, 16384]
forward: 4.61ms -> 4.80ms
backward: 26.30ms -> 27.10ms

Test Plan: buck test mode/dev-nosan //caffe2/test:nn -- "LayerNorm"

Differential Revision: D30861431